# documentation/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SeriesBookIn": {
        "description": "Привязка книги к циклу: либо существующая книга (`book_id`), либо новая\nпо названию — тогда заводим запись в КАТАЛОГЕ, не кладя на полку. Так в цикле\nпоявляются будущие книги («что дальше»): они есть в каталоге, но не у читателя.\n`cover_url`/`external_id` приходят из поиска (Google Books) — по ним книга\nполучает обложку и фоновое обогащение.",
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "book_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Book Id"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "external_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Id"
          },
          "series_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Series Index"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "SeriesBookIn",
        "type": "object"
      },
      "SeriesCreate": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "default": "reading",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "SeriesCreate",
        "type": "object"
      },
      "SeriesUpdate": {
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "title": "SeriesUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Персональная библиотека для атмосферных литературных вечеров. Интерактивная документация: /docs (Swagger) и /redoc.",
    "title": "nocturne API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/series": {
      "get": {
        "description": "Полка циклов: читаю → прочитано → перестала читать.",
        "operationId": "read_series_api_v1_series_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Series",
        "tags": [
          "series"
        ]
      },
      "post": {
        "description": "Создать цикл. Сразу кладём его на полку пользователя со статусом.\nЦикл — общая сущность → создание только admin (з.90а).",
        "operationId": "create_series_api_v1_series_post",
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Series",
        "tags": [
          "series"
        ]
      }
    },
    "/api/v1/series/{series_id}": {
      "delete": {
        "description": "Удалить цикл. Книги остаются в каталоге — просто теряют привязку.\nЦикл общий → удаление только admin (з.90а).",
        "operationId": "delete_series_api_v1_series__series_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Series",
        "tags": [
          "series"
        ]
      },
      "get": {
        "operationId": "read_one_api_v1_series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read One",
        "tags": [
          "series"
        ]
      },
      "patch": {
        "description": "Правка цикла и/или смена статуса.\n\nРазделение прав (з.90а): цикл — ОБЩАЯ сущность (как книга), поэтому правка\nназвания/автора/описания меняет её для всех → только admin. А статус\n(«Читаю» / «Прочитан» / «Брошен») — личное действие читателя, доступен всем.",
        "operationId": "update_series_api_v1_series__series_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Series",
        "tags": [
          "series"
        ]
      }
    },
    "/api/v1/series/{series_id}/books": {
      "post": {
        "description": "Привязать книгу к циклу (со страницы цикла).\nКнига может отсутствовать на полке пользователя — тогда в цикле она\nпоказывается как «что дальше».\nСостав цикла — общие данные → только admin (з.90а).",
        "operationId": "add_book_to_series_api_v1_series__series_id__books_post",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesBookIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Book To Series",
        "tags": [
          "series"
        ]
      }
    },
    "/api/v1/series/{series_id}/books/{book_id}": {
      "delete": {
        "operationId": "remove_book_from_series_api_v1_series__series_id__books__book_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "title": "Book Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Book From Series",
        "tags": [
          "series"
        ]
      }
    },
    "/api/v1/series/{series_id}/design": {
      "post": {
        "description": "Сгенерировать экслибрис цикла (задача 89). Тратит токены → только admin,\nпо кнопке. Опирается на описание цикла — поэтому имеет смысл сначала его\nзаполнить.\nСессию держим КОРОТКО: между чтением и записью идёт долгий AI-вызов.",
        "operationId": "generate_design_api_v1_series__series_id__design_post",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "ru",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Design",
        "tags": [
          "series"
        ]
      }
    },
...
  }
}